more refactoring, add dedicated tenant cleanup struct
