+++ fleetshard/pkg/central/reconciler/reconciler_test.go
@@ -547,32 +547,6 @@ func TestReconcileDelete(t *testing.T) {
 	assert.True(t, k8sErrors.IsNotFound(err))
 }
 
-func TestDisablePauseAnnotation(t *testing.T) {
-	fakeClient, _, r := getClientTrackerAndReconciler(
-		t,
-		defaultCentralConfig,
-		nil,
-		useRoutesReconcilerOptions,
-	)
-
-	_, err := r.Reconcile(context.TODO(), simpleManagedCentral)
-	require.NoError(t, err)
-
-	central := &v1alpha1.Central{}
-	err = fakeClient.Get(context.TODO(), client.ObjectKey{Name: centralName, Namespace: centralNamespace}, central)
-	require.NoError(t, err)
-	central.Annotations[PauseReconcileAnnotation] = "true"
-	err = fakeClient.Update(context.TODO(), central)
-	require.NoError(t, err)
-
-	err = r.disablePauseReconcileIfPresent(context.TODO(), central)
-	require.NoError(t, err)
-
-	err = fakeClient.Get(context.TODO(), client.ObjectKey{Name: centralName, Namespace: centralNamespace}, central)
-	require.NoError(t, err)
-	require.Equal(t, "false", central.Annotations[PauseReconcileAnnotation])
-}
-
 func TestReconcileDeleteWithManagedDB(t *testing.T) {
 	managedDBProvisioningClient := &cloudprovider.DBClientMock{}
 	managedDBProvisioningClient.EnsureDBProvisionedFunc = func(_ context.Context, databaseID, acsInstanceID, _ string, _ bool) error {
@@ -2331,7 +2305,7 @@ func TestReconciler_reconcileNamespace(t *testing.T) {
 					return client.Create(ctx, obj, opts...)
 				},
 			})
-			err := r.namespaceReconciler.ReconcileNamespace(context.Background(), getDesiredNamespace(managedCentral))
+			err := r.namespaceReconciler.Reconcile(context.Background(), getDesiredNamespace(managedCentral))
 			if tt.wantErr {
 				assert.Error(t, err)
 			} else {
